requirement.txt
pycryptodomex==3.14.1
pyinstaller==5.1